requirements.txt
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0

# Daedalus Phase 1 additions
nest_asyncio>=1.6.0
upstash-redis>=1.0.0
motor>=3.0.0
pymongo>=4.0.0
pyyaml>=6.0.0
duckduckgo-search>=5.0.0
aiohttp>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0